backend/requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.1
geoalchemy2==0.15.2

# Task Queue
celery[redis]==5.4.0
redis==5.2.1

# HTTP / Data Fetching
httpx==0.28.1
aiohttp==3.11.11
feedparser==6.0.11

# Data Processing
pandas==2.2.3
numpy==2.2.1

# LLM / RAG Pipeline
chromadb==0.5.23
sentence-transformers==3.3.1
langchain==0.3.14
langchain-community==0.3.14

# NLP / Text Processing
gdeltPyR==0.1.10

# Geospatial
shapely==2.0.6
pyproj==3.7.0

# Utilities
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
python-dateutil==2.9.0.post0
structlog==24.4.0
tenacity==9.0.0
# Authentication
bcrypt==4.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18
PyJWT==2.10.1
stripe
email-validator
